Add Start Screen Component
Now, instead of putting the user on the form upon load, we let the user choose whether to create, continue, or load a sample resume. The start screen has a menu pane where users can choose how to start the application; a switch conditional branch determines what sort of data is loaded in startApp() in App.js.